DownloadThread: customizeImage: Setup image customization for Zynq platform
This patch will add GUI elements for selecting the project to be
configured. After the image write step is done, the SD card will be mounted
and the BOOT partition will be configured with the selected project files.
This includes copying the BOOT.BIN, device tree blob and Linux Kernel in
the BOOT directory.

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>